+++ native/src/main/java/io/ballerina/stdlib/sql/nativeimpl/OutParameterProcessor.java
@@ -18,6 +18,7 @@
 
 package io.ballerina.stdlib.sql.nativeimpl;
 
+import io.ballerina.runtime.api.TypeTags;
 import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BTypedesc;
@@ -55,16 +56,23 @@ public class OutParameterProcessor {
     private OutParameterProcessor() {
     }
 
-    public static Object get(BObject result, BTypedesc typeDesc) {
-        return get(result, typeDesc, DefaultResultParameterProcessor.getInstance());
+    public static Object getOutParameterValue(BObject result, BTypedesc typeDesc) {
+        return get(result, typeDesc, DefaultResultParameterProcessor.getInstance(), "OutParameter");
     }
 
-    public static Object get(
-            BObject result, BTypedesc typeDesc, AbstractResultParameterProcessor resultParameterProcessor) {
+    public static Object getInOutParameterValue(BObject result, BTypedesc typeDesc) {
+        return get(result, typeDesc, DefaultResultParameterProcessor.getInstance(), "InOutParameter");
+    }
+
+    public static Object get(BObject result, BTypedesc typeDesc,
+                              AbstractResultParameterProcessor resultParameterProcessor, String parameterType) {
         int sqlType = (int) result.getNativeData(Constants.ParameterObject.SQL_TYPE_NATIVE_DATA);
         Object value = result.getNativeData(Constants.ParameterObject.VALUE_NATIVE_DATA);
         Type ballerinaType = typeDesc.getDescribingType();
         try {
+            if (ballerinaType.getTag() == TypeTags.UNION_TAG) {
+                throw new ApplicationError(parameterType + " 'get' function does not support union return type.");
+            }
             switch (sqlType) {
                 case Types.CHAR:
                 case Types.VARCHAR: