+++ changelog.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- [Tooling support for SQL client](https://github.com/ballerina-platform/ballerina-standard-library/issues/2058)
 
 ### Changed
 
 - [Accept escaped backtick as insertions in parameterised query](https://github.com/ballerina-platform/ballerina-standard-library/issues/2056)
 - [Add support for handling union types in queryRow()](https://github.com/ballerina-platform/ballerina-standard-library/issues/2333)
 - [Validate connection pool configurations](https://github.com/ballerina-platform/ballerina-standard-library/issues/2355)
+- [Change queryRow return type to anydata](https://github.com/ballerina-platform/ballerina-standard-library/issues/2390)
+- [Make OutParameter get function parameter optional](https://github.com/ballerina-platform/ballerina-standard-library/issues/2388)
 
 ## [1.0.0] - 2021-10-09